Enum fdeflate::DecompressionError

pub enum DecompressionError {

    BadZlibHeader,
    InsufficientInput,
    InvalidBlockType,
    InvalidUncompressedBlockLength,
    InvalidHlit,
    InvalidHdist,
    InvalidCodeLengthRepeat,
    BadCodeLengthHuffmanTree,
    BadLiteralLengthHuffmanTree,
    BadDistanceHuffmanTree,
    InvalidLiteralLengthCode,
    InvalidDistanceCode,
    InputStartsWithRun,
    DistanceTooFarBack,
    WrongChecksum,
    ExtraInput,

}

An error encountered while decompressing a deflate stream.

Variants

BadZlibHeader

The zlib header is corrupt.

InsufficientInput

All input was consumed, but the end of the stream hasn’t been reached.

InvalidBlockType

A block header specifies an invalid block type.

InvalidUncompressedBlockLength

An uncompressed block’s NLEN value is invalid.

InvalidHlit

Too many literals were specified.

InvalidHdist

Too many distance codes were specified.

InvalidCodeLengthRepeat

Attempted to repeat a previous code before reading any codes, or past the end of the code lengths.

BadCodeLengthHuffmanTree

The stream doesn’t specify a valid huffman tree.

BadLiteralLengthHuffmanTree

The stream doesn’t specify a valid huffman tree.

BadDistanceHuffmanTree

The stream doesn’t specify a valid huffman tree.

InvalidLiteralLengthCode

The stream contains a literal/length code that was not allowed by the header.

InvalidDistanceCode

The stream contains a distance code that was not allowed by the header.

InputStartsWithRun

The stream contains contains back-reference as the first symbol.

DistanceTooFarBack

The stream contains a back-reference that is too far back.

WrongChecksum

The deflate stream checksum is incorrect.

ExtraInput

Extra input data.