fdeflate

Module tables

Constants

BITMASKS 🔒

CLCL_ORDER 🔒

Order of the length code length alphabet (derived from deflate spec).

DISTANCE_TABLE_ENTRIES 🔒

The distance table is a 512-entry table that maps 9 bits of distance symbols to their meaning.

DIST_SYM_TO_DIST_BASE 🔒

The base distance for each distance code (derived from deflate spec).

DIST_SYM_TO_DIST_EXTRA 🔒

Number of extra bits for each distance code (derived from deflate spec.)

FIXED_DIST_TABLE 🔒

FIXED_LITLEN_TABLE 🔒

HUFFMAN_CODES 🔒

HUFFMAN_LENGTHS 🔒

Hard-coded Huffman codes used regardless of the input.

LENGTH_TO_LEN_EXTRA 🔒

Number of extra bits for length values (derived from deflate spec).

LENGTH_TO_SYMBOL 🔒

Length code for length values (derived from deflate spec).

LEN_SYM_TO_LEN_BASE 🔒

The base length for each length code (derived from deflate spec).

LEN_SYM_TO_LEN_EXTRA 🔒

Number of extra bits for each length code (derived from deflate spec).

LITLEN_TABLE_ENTRIES 🔒

The main litlen_table uses a 12-bit input to lookup the meaning of the symbol. The table is split into 4 sections: