Module fdeflate::decompress

Structs

BlockHeader 🔒

CompressedBlock 🔒

The Decompressor state for a compressed block.

Decompressor

Decompressor for arbitrary zlib streams.

Enums

BoundedDecompressionError

An error encountered while decompressing a deflate stream given a bounded maximum output.

DecompressionError

An error encountered while decompressing a deflate stream.

State 🔒

Constants

EXCEPTIONAL_ENTRY 🔒

FDEFLATE_COMPRESSED_BLOCK 🔒

LITERAL_ENTRY 🔒

SECONDARY_TABLE_ENTRY 🔒

Functions

decompress_to_vec

Decompress the given data.

decompress_to_vec_bounded

Decompress the given data, returning an error if the output is larger than maxlen bytes.