Module fdeflate::tables

Constants

BITMASKS 🔒

CLCL_ORDER 🔒

Order of the length code length alphabet (derived from deflate spec).

DIST_SYM_TO_DIST_BASE 🔒

The base distance for each distance code (derived from deflate spec).

DIST_SYM_TO_DIST_EXTRA 🔒

Number of extra bits for each distance code (derived from deflate spec.)

FDEFLATE_DIST_DECODE_TABLE 🔒

FDEFLATE_LITLEN_DECODE_TABLE 🔒

FIXED_CODE_LENGTHS 🔒

HUFFMAN_CODES 🔒

HUFFMAN_LENGTHS 🔒

Hard-coded Huffman codes used regardless of the input.

LENGTH_TO_LEN_EXTRA 🔒

Number of extra bits for length values (derived from deflate spec).

LENGTH_TO_SYMBOL 🔒

Length code for length values (derived from deflate spec).

LEN_SYM_TO_LEN_BASE 🔒

The base length for each length code (derived from deflate spec).

LEN_SYM_TO_LEN_EXTRA 🔒

Number of extra bits for each length code (derived from deflate spec).

Functions

make_fixed_code_lengths 🔒